chore: rename request config
