+++ packages/orchestration/src/util/index.ts
@@ -1,3 +1,3 @@
 export * from './filtering.js';
 export * from './grounding.js';
-export * from './request-config.js';
+export * from './orchestration-request-config.js';
